move some tar helpers from server to unpack
