+++ cmd/gitserver/server/vcs_syncer_go_modules.go
@@ -104,7 +104,7 @@ func unzip(mod module.Version, zipBytes []byte, workDir string) error {
 	err = unpack.Zip(br, int64(br.Len()), workDir, unpack.Opts{
 		SkipInvalid: true,
 		Filter: func(path string, file fs.FileInfo) bool {
-			_, malicious := isPotentiallyMaliciousFilepathInArchive(path, workDir)
+			_, malicious := unpack.IsPotentiallyMaliciousFilepathInArchive(path, workDir)
 			_, ok := valid[path]
 			return ok && !malicious
 		},
